# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Top All Money API",
        "description": "Documentation de l'API top all money",
        "contact": {
            "email": "support@tonapp.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local"
        },
        {
            "url": "https://api.tonapp.com",
            "description": "Production"
        }
    ],
    "paths": {
...
        "/api/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Lister les catégories",
                "operationId": "ac9c4d978ea4e5b72915ca9728e5b8b1",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "isDirty",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ]
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "updatedAt:desc"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 200,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Créer une catégorie",
                "operationId": "cca49a87d9b0ebb9a98ad7c154b71737",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Créé"
                    }
                }
            }
        },
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Afficher une catégorie",
                "operationId": "09767bb2e47dde77dd1e2591428b0ae1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Mettre à jour une catégorie",
                "operationId": "7dcf71887e585de7f881e9eb98d86126",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Soft delete",
                "operationId": "9b61b65823f4a21f182506b34ddf805c",
                "responses": {
                    "200": {
                        "description": "Supprimé"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Categories"
                ],
                "summary": "Modifier partiellement une catégorie",
                "operationId": "b13d20619d95750562d5ccb8ae81d765",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
        "/api/categories/bulk": {
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Upsert en masse",
                "operationId": "1ed85f2a8d7548bb59b66fe93fc7ca7d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "items"
                                ],
                                "properties": {
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/CategoryUpdateRequest"
                                                },
                                                {
                                                    "properties": {
                                                        "id": {
                                                            "type": "string",
                                                            "format": "uuid",
                                                            "nullable": true
                                                        },
                                                        "type": {
                                                            "type": "string",
                                                            "enum": [
                                                                "CREATE",
                                                                "UPDATE",
                                                                "DELETE"
                                                            ]
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            ]
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/categories/import": {
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Importer un CSV",
                "operationId": "5d1e90c33612f4739181d19d4364a101",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Import terminé"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CategoryCreateRequest": {
                "properties": {
                    "code": {
                        "type": "string",
                        "maxLength": 255,
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "typeEntry": {
                        "type": "string",
                        "enum": [
                            "DEBIT",
                            "CREDIT"
                        ],
                        "nullable": true
                    },
                    "account": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "CategoryUpdateRequest": {
                "properties": {
                    "code": {
                        "type": "string",
                        "maxLength": 255,
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "typeEntry": {
                        "type": "string",
                        "enum": [
                            "DEBIT",
                            "CREDIT"
                        ],
                        "nullable": true
                    },
                    "account": {
                        "type": "string",
                        "nullable": true
                    },
                    "isDirty": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}